# task 4/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScanSummary {
  id: number;
  label: string;
  score: number;
  grade: string;
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  os?: string | null;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  totalFindings: number;
  createdAt: string;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
  Info: 'Info',
} as const;

export interface Finding {
  id: number;
  title: string;
  description: string;
  severity: FindingSeverity;
  category: string;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  cvssScore?: number | null;
}

export interface SystemInfo {
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  architecture?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  ram?: string | null;
  /** @nullable */
  uptime?: string | null;
  /** @nullable */
  domain?: string | null;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  biosVersion?: string | null;
}

export interface ScanDetail {
  id: number;
  label: string;
  score: number;
  grade: string;
  systemInfo?: SystemInfo;
  findings: Finding[];
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  totalFindings: number;
  createdAt: string;
}

export type ScanIngestFindingsItem = {
  title: string;
  description: string;
  severity: string;
  category: string;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  cvssScore?: number | null;
};

export interface ScanIngest {
  label: string;
  systemInfo?: SystemInfo;
  findings: ScanIngestFindingsItem[];
}

export interface IngestResult {
  scanId: number;
  score: number;
  grade: string;
  findingCount: number;
}

export interface ScoreTrend {
  date: string;
  score: number;
  label: string;
}

export type DashboardStatsTopCategoriesItem = {
  category: string;
  count: number;
};

export interface DashboardStats {
  totalScans: number;
  /** @nullable */
  averageScore?: number | null;
  /** @nullable */
  lastScanAt?: string | null;
  totalCritical: number;
  totalHigh: number;
  topCategories: DashboardStatsTopCategoriesItem[];
  scoreTrend: ScoreTrend[];
}

export type ListScansParams = {
limit?: number;
offset?: number;
};