# task 4/lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsTopCategoriesItem } from './dashboardStatsTopCategoriesItem';
import type { ScoreTrend } from './scoreTrend';

export interface DashboardStats {
  totalScans: number;
  /** @nullable */
  averageScore?: number | null;
  /** @nullable */
  lastScanAt?: Date | null;
  totalCritical: number;
  totalHigh: number;
  topCategories: DashboardStatsTopCategoriesItem[];
  scoreTrend: ScoreTrend[];
}