# task 4/lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';

export interface Finding {
  id: number;
  title: string;
  description: string;
  severity: FindingSeverity;
  category: string;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  cvssScore?: number | null;
}