# task 4/lib/api-zod/src/generated/types/ingestResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IngestResult {
  scanId: number;
  score: number;
  grade: string;
  findingCount: number;
}