# task 4/lib/api-zod/src/generated/types/scanDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { SystemInfo } from './systemInfo';

export interface ScanDetail {
  id: number;
  label: string;
  score: number;
  grade: string;
  systemInfo?: SystemInfo;
  findings: Finding[];
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  totalFindings: number;
  createdAt: Date;
}