# task 4/lib/api-zod/src/generated/types/scanIngest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanIngestFindingsItem } from './scanIngestFindingsItem';
import type { SystemInfo } from './systemInfo';

export interface ScanIngest {
  label: string;
  systemInfo?: SystemInfo;
  findings: ScanIngestFindingsItem[];
}