# task 4/lib/api-zod/src/generated/types/scanSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanSummary {
  id: number;
  label: string;
  score: number;
  grade: string;
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  os?: string | null;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  totalFindings: number;
  createdAt: Date;
}