# task 4/lib/api-zod/src/generated/types/systemInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberAudit Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SystemInfo {
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  architecture?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  ram?: string | null;
  /** @nullable */
  uptime?: string | null;
  /** @nullable */
  domain?: string | null;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  biosVersion?: string | null;
}